refactor: remove @tanstack/react-table dependency and introduce AggregationCard and ReachHeroSection components for improved data visualization in analyze page
