+++ packages/shared/package.json
@@ -113,7 +113,6 @@
     "@radix-ui/react-popover": "^1.1.15",
     "@radix-ui/react-slider": "^1.3.6",
     "@radix-ui/react-tooltip": "^1.2.8",
-    "@tanstack/react-table": "^8.21.3",
     "@tippyjs/react": "^4.2.6",
     "check-password-strength": "^2.0.10",
     "fetch-event-stream": "^0.1.5",
