+++ packages/shared/src/features/opportunity/context/PendingSubmissionContext.tsx
@@ -1,5 +1,5 @@
-import type { ReactNode } from 'react';
-import React, { createContext, useContext, useState, useCallback } from 'react';
+import { useState, useCallback } from 'react';
+import { createContextProvider } from '@kickass-coderz/react';
 
 export interface PendingSubmissionUrl {
   type: 'url';
@@ -13,50 +13,31 @@ export interface PendingSubmissionFile {
 
 export type PendingSubmission = PendingSubmissionUrl | PendingSubmissionFile;
 
-interface PendingSubmissionContextType {
-  pendingSubmission: PendingSubmission | null;
-  setPendingSubmission: (submission: PendingSubmission) => void;
-  clearPendingSubmission: () => void;
-}
+const [PendingSubmissionProvider, usePendingSubmission] = createContextProvider(
+  () => {
+    const [pendingSubmission, setPendingSubmissionState] =
+      useState<PendingSubmission | null>(null);
 
-const PendingSubmissionContext =
-  createContext<PendingSubmissionContextType | null>(null);
-
-export const PendingSubmissionProvider = ({
-  children,
-}: {
-  children: ReactNode;
-}) => {
-  const [pendingSubmission, setPendingSubmissionState] =
-    useState<PendingSubmission | null>(null);
-
-  const setPendingSubmission = useCallback((submission: PendingSubmission) => {
-    setPendingSubmissionState(submission);
-  }, []);
-
-  const clearPendingSubmission = useCallback(() => {
-    setPendingSubmissionState(null);
-  }, []);
-
-  return (
-    <PendingSubmissionContext.Provider
-      value={{
-        pendingSubmission,
-        setPendingSubmission,
-        clearPendingSubmission,
-      }}
-    >
-      {children}
-    </PendingSubmissionContext.Provider>
-  );
-};
-
-export const usePendingSubmission = (): PendingSubmissionContextType => {
-  const context = useContext(PendingSubmissionContext);
-  if (!context) {
-    throw new Error(
-      'usePendingSubmission must be used within PendingSubmissionProvider',
+    const setPendingSubmission = useCallback(
+      (submission: PendingSubmission) => {
+        setPendingSubmissionState(submission);
+      },
+      [],
     );
-  }
-  return context;
-};
+
+    const clearPendingSubmission = useCallback(() => {
+      setPendingSubmissionState(null);
+    }, []);
+
+    return {
+      pendingSubmission,
+      setPendingSubmission,
+      clearPendingSubmission,
+    };
+  },
+  {
+    scope: 'PendingSubmissionProvider',
+  },
+);
+
+export { PendingSubmissionProvider, usePendingSubmission };
